+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [3.0.7-beta] — 2026-TBD
+- Added 'is_slug' param for text inputs to force slug format - ADDED
+- Package loader is not respecting load order due to AI slop - FIXED
+
 ## [3.0.6-beta] — 2026-04-30
 - Field builder refactor to new standards - CHANGED
 - ayecode_sd_register() function added to register blocks for better efficiency - ADDED